+++ src/main/java/insane96mcp/iguanatweaksexpanded/data/generator/ITERecipeProvider.java
@@ -1,7 +1,6 @@
 package insane96mcp.iguanatweaksexpanded.data.generator;
 
 import insane96mcp.iguanatweaksexpanded.IguanaTweaksExpanded;
-import insane96mcp.iguanatweaksexpanded.module.experience.enchanting.EnchantingFeature;
 import insane96mcp.iguanatweaksexpanded.module.hungerhealth.fooddrinks.FoodDrinks;
 import insane96mcp.iguanatweaksexpanded.module.items.altimeter.Altimeter;
 import insane96mcp.iguanatweaksexpanded.module.items.copper.CopperExpansion;
@@ -250,14 +249,6 @@ public class ITERecipeProvider extends RecipeProvider implements IConditionBuild
                 .define('i', Items.IRON_INGOT)
                 .define('r', Items.REDSTONE)
                 .unlockedBy("has_durium_ingot", has(Durium.INGOT.get()))
-                .save(writer);
-
-        ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, EnchantingFeature.ANCIENT_LAPIS.get(), 1)
-                .requires(EnchantingFeature.CLEANSED_LAPIS.get(), 1)
-                .requires(Items.NETHERITE_SCRAP, 1)
-                .requires(Items.EXPERIENCE_BOTTLE, 1)
-                .unlockedBy("has_netherite_scrap", has(Items.NETHERITE_SCRAP))
-                .unlockedBy("has_cleansed_lapis", has(EnchantingFeature.CLEANSED_LAPIS.get()))
                 .save(writer);
 
         //Solarium ball and forging