+++ changelog.md
@@ -6,6 +6,8 @@
   * Blood pact: mending but the durability bar is your health (Treasure)
   * Soulbound: items are kept in the inventory on death (Treasure)
   * Adrenaline: mining and attack speed increases as the durability decreases
+* Removed Ancient lapis
+  * The mechanic has been moved to enchanted Cleansed Lapis, made with a Cleansed Lapis and an Experience bottle in an Anvil
 * Forge Hammers
   * Movement Speed -15% -> -10%
   * Entity Reach 1.5 -> 2